[Fix] 파일 복구
